# src/main/resources/api.yaml
openapi: 3.0.4
info:
  title: Task Manager API
  description: REST API для управления задачами
  contact:
    email: hdnhdxv@gmail.com
  version: 0.1.0
...
paths:
  /register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      security: []
      description: Создает нового пользователя в системе
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: user123
              password: SecurePass123!
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Неверные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Пользователь с таким username уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /login:
    post:
      tags:
        - Authentication
      summary: Вход в систему
      security: []
      description: Аутентификация пользователя и получение токенов
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: exampleUsername
              password: SecurePass123!
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseToken'
        '400':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизован - неверный username или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен - пользователь заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Слишком много попыток входа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Вставьте JWT токен в формате: `Bearer <token>`
        Получить токен можно через `/login` или `/token` endpoints.
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
          description: Пароль (мин. 8 символов)
          minLength: 8
          maxLength: 100
          example: SecurePass123!
    # ============ RESPONSE SCHEMAS ============
    AuthResponseToken:
      type: object
      properties:
        token:
          type: string
          description: Access токен для аутентификации запросов
        user:
          $ref: '#/components/schemas/UserResponse'
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        token:
          type: string
          description: Access токен для аутентификации запросов
    UserResponse:
      type: object
      properties:
        id:
          type: string
          description: Уникальный идентификатор пользователя
        username:
          type: string
          description: username пользователя
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
          description: Пароль пользователя
          example: SecurePass123!
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Тип ошибки
        message:
          type: string
          description: Сообщение об ошибке
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
security:
  - BearerAuth: [ ]